requirements.txt
openai
datasets
tqdm
spacy
python-dotenv
sentence-transformers
scikit-learn
numpy
phonenumbers
streamlit
pandas 
matplotlib 
seaborn 
plotly
